Refactor storage functionality into base class, in preparation for a pg version
